+++ playwright/set1-tests.spec.ts
@@ -1,4 +1,6 @@
 import {test} from '@playwright/test';
-import createAdvancedTaskControlsTests from './advanced-task-controls-test.spec';
 
-test.describe('Advanced Task Controls Tests', createAdvancedTaskControlsTests);
+import createIncomingTaskAndControlsMultiSessionTests from './tests/incoming-task-and-controls-multi-session.spec';
+import createDigitalIncomingTaskAndTaskControlsTests from './tests/digital-incoming-task-and-task-controls.spec';
+test.describe('Digital Incoming and Task Controls Tests', createDigitalIncomingTaskAndTaskControlsTests);
+test.describe('Incoming Task Multi-Session Tests', createIncomingTaskAndControlsMultiSessionTests);
