+++ playwright/set2-tests.spec.ts
@@ -1,6 +1,3 @@
 import {test} from '@playwright/test';
-import createBasicTaskControlsTests from './basic-task-controls-test.spec';
-import createAdvanceCombinationsTests from './advanceCombinationsTests.spec';
-
-test.describe('Basic Task Controls Tests', createBasicTaskControlsTests);
+import createAdvanceCombinationsTests from './tests/advance-task-control-combinations-test.spec';
 test.describe('Advanced Combinations Tests', createAdvanceCombinationsTests);
