+++ playwright/set4-tests.spec.ts
@@ -1,7 +1,3 @@
 import {test} from '@playwright/test';
-
-import createIncomingTaskTests from './incoming-task-test.spec';
-import createIncomingTaskMultiSessionTests from './incoming-task-multi-session.spec';
-
-test.describe('Incoming Task Tests', createIncomingTaskTests);
-test.describe('Incoming Task Multi-Session Tests', createIncomingTaskMultiSessionTests);
+import createCallTaskControlsTests from './tests/basic-task-controls-test.spec';
+test.describe('Call Task Controls Tests', createCallTaskControlsTests);
