+++ playwright/set5-tests.spec.ts
@@ -1,0 +1,4 @@
+import {test} from '@playwright/test';
+import createAdvancedTaskControlsTests from './tests/advanced-task-controls-test.spec';
+
+test.describe('Advanced Task Controls Tests', createAdvancedTaskControlsTests);
