+++ playwright/test-data.ts
@@ -5,41 +5,52 @@ require('dotenv').config();
 export const USER_SETS = {
   SET_1: {
     AGENTS: {
-      AGENT1: {username: 'User16', extension: '1016', agentName: 'User16 Agent16'},
-      AGENT2: {username: 'user9', extension: '1009', agentName: 'User9 Agent9'},
+      AGENT1: {username: 'User15', extension: '1015', agentName: 'User15 Agent15'},
+      AGENT2: {username: 'user16', extension: '1016', agentName: 'User16 Agent16'},
     },
-    QUEUE_NAME: 'Queue 5',
+    QUEUE_NAME: 'Queue e2e 1',
+    CHAT_URL: 'https://widgets.webex.com/chat-client-e2e.html',
+    EMAIL_ENTRY_POINT: 'ccsdk.wbx.ai.e2e@gmail.com',
+    DIAL_NUMBER: env.PW_DIAL_NUMBER1,
+  },
+  SET_2: {
+    AGENTS: {
+      AGENT1: {username: 'user17', extension: '1017', agentName: 'User17 Agent17'},
+      AGENT2: {username: 'user18', extension: '1018', agentName: 'User18 Agent18'},
+    },
+    QUEUE_NAME: 'Queue e2e 2',
+    CHAT_URL: 'https://widgets.webex.com/chat-client-e2e-2.html',
+    EMAIL_ENTRY_POINT: '',
+    DIAL_NUMBER: env.PW_DIAL_NUMBER2,
+  },
+  SET_3: {
+    AGENTS: {
+      AGENT1: {username: 'user19', extension: '1019', agentName: 'User19 Agent19'},
+      AGENT2: {username: 'user20', extension: '1020', agentName: 'User20 Agent20'},
+    },
+    QUEUE_NAME: 'Queue e2e 3',
     CHAT_URL: '',
     EMAIL_ENTRY_POINT: '',
     DIAL_NUMBER: env.PW_DIAL_NUMBER3,
   },
-  SET_2: {
+  SET_4: {
     AGENTS: {
-      AGENT1: {username: 'user1', extension: '1001', agentName: 'User1 Agent1'},
-      AGENT2: {username: 'user2', extension: '1002', agentName: 'User2 Agent2'},
+      AGENT1: {username: 'user21', extension: '1021', agentName: 'User21 Agent21'},
+      AGENT2: {username: 'user22', extension: '1022', agentName: 'User22 Agent22'},
     },
-    QUEUE_NAME: 'Queue-1',
-    CHAT_URL: 'https://widgets.webex.com/chat-client.html',
-    EMAIL_ENTRY_POINT: 'ccsdk.wbx.ai@gmail.com',
-    DIAL_NUMBER: env.PW_DIAL_NUMBER1,
-  },
-  SET_3: {
-    AGENTS: {
-      AGENT1: {username: 'user15', extension: '1015', agentName: 'User15 Agent15'},
-    },
-    QUEUE_NAME: '',
+    QUEUE_NAME: 'Queue e2e 4',
     CHAT_URL: '',
     EMAIL_ENTRY_POINT: '',
     DIAL_NUMBER: env.PW_DIAL_NUMBER4,
   },
-  SET_4: {
+  SET_5: {
     AGENTS: {
-      AGENT1: {username: 'user11', extension: '1011', agentName: 'User11 Agent11'},
-      AGENT2: {username: 'user13', extension: '1013', agentName: 'User13 Agent13'},
+      AGENT1: {username: 'user23', extension: '1023', agentName: 'User23 Agent23'},
+      AGENT2: {username: 'user24', extension: '1024', agentName: 'User24 Agent24'},
     },
-    QUEUE_NAME: 'Queue 4',
-    CHAT_URL: 'https://widgets.webex.com/chat-client-e2e.html',
-    EMAIL_ENTRY_POINT: 'ccsdk.wbx.ai.e2e@gmail.com',
-    DIAL_NUMBER: env.PW_DIAL_NUMBER2,
+    QUEUE_NAME: 'Queue e2e 5',
+    CHAT_URL: '',
+    EMAIL_ENTRY_POINT: '',
+    DIAL_NUMBER: env.PW_DIAL_NUMBER5,
   },
 };
